meta-aspeed-sdk: apply firmware update immediately
Currently, the webui-vue uses multipart/form-data as the Content-Type
and does not set the applyTime field. As a result, the update service
defaults applyTime to OnReset,  which prevents the BMC from rebooting
immediately after a firmware update via the WebUI.

Change update service defaults applyTime to "Immediate". This ensures
the BMC reboots immediately after a firmware update via the WebUI,
instead of waiting for a reset.

Signed-off-by: Vince Chang <vince_chang@aspeedtech.com>
Change-Id: Ia1be6170e754fe7b2d5f7ababab8e1a0296088bb