meta-aspeed-sdk: ast2600: move and rename secure CoT settings to clarify RSA-only support
In the previous design, the shared "aspeed-secure-cot.inc" file was used for
all ASPEED SoCs to support Chain of Trust (CoT) configurations. However, this
approach introduced confusion because the AST2600 platform does not support
ECDSA, while AST2700 defaults to ECDSA as its CoT algorithm.

To avoid misleading users, the RSA-only configuration for AST2600 is now
renamed to ast2600-secure-cot.inc and relocated to the ast2600-specific
meta-layer. This change clarifies that the file only supports RSA-based secure
boot settings and prevents unintended use with platforms that expect ECDSA
support.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: Ic1d3087d3876e5b434ba5908ed8f22cab7bf29e4