meta-aspeed-sdk:ast2700-sdk: use ??= for defaults, ?= for machine overrides
In ast2700-sdk.inc, switch PREFERRED_PROVIDER_virtual/* to ??= to provide fallback defaults at the SDK layer.
In ast2700-default-irot.conf, change assignments to ?= so machine configs can supply defaults but remain overridable.

Set virtual/provider ??= "A" in ast2700-sdk.inc so that the SDK layer provides A as the fallback default.
Then, set virtual/provider ?= "B" in the machine config, so the machine can supply B as its default if nothing else is chosen.
Finally, the user can set virtual/provider = "C" in local.conf. **Because local.conf is parsed before the machine config and only = can be used there**,
this hard assignment will take precedence over both ??= and ?=.

Reference:
https://docs.yoctoproject.org/bitbake/2.12/index.html

1. Setting a default value (?=)
A ?= "aval"
If A is set at the time this statement is parsed, the variable retains its value. However, if A is not set, the variable is set to “aval”.
This assignment is immediate. Consequently, if multiple “?=” assignments to a single variable exist, the first of those ends up getting used.

2. Setting a weak default value (??=)
The weak default value of a variable is the value which that variable will expand to if no value has been assigned to it via any of the other assignment operators.
The "??=" operator takes effect immediately, replacing any previously defined weak default value.

??= gives the weakest fallback value, and it can be overridden by either `?=` or `=`.
Using ??= multiple times will update the weak default.

The priority is `=` > `?=` > `??=`.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I37b0a95f80a95272e76053206975aae10b5dcdee