meta-aspeed-sdk:aspeed-g7: Remove bootmcu-spl bbappend as U-Boot FIT signing is no longer required
This bbappend was originally used to install the unsigned SPL binaries so that
the U-Boot recipe could sign its FIT image and embed the public key in
u-boot-spl.dtb to support Chain of Trust (CoT).

However, starting from AST2700 A1, the U-Boot FIT image has been replaced by a
SoC manifest, which is now signed using the Caliptra signing tool instead of
U-Boot's signing mechanism. As a result, the U-Boot FIT signing process and the
use of unsigned SPL artifacts are no longer required.

This commit removes the bootmcu-spl_%.bbappend file since it is obsolete.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I0df81de77607f0caf7a42f8e0de98ff1c90b9012