meta-aspeed-sdk: Refactor fmc-images dependency handling and deploy ordering
In the previous design, "fmc-images" was added to the DEPENDS variable so that
its do_populate_sysroot task would run before dependent recipes executed their
do_configure tasks. However, these recipes actually require "fmc-images" to be
completed at the do_deploy stage instead.

Because do_deploy of "fmc-images" is not guaranteed to execute before
do_populate_sysroot (it is only set to run before do_build and after
do_compile), this could cause build ordering issues. To fix this, explicit
task-level dependencies are added on "fmc-images:do_deploy" in the specific
tasks that need it.

The "virtual/bootloader" remains in the DEPENDS variable in
"aspeed-image-gen-secureboot.bb" because it requires "u-boot-nodtb.bin",
which is installed during the do_install task rather than do_deploy.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: Idb3e61d6f189684ae06019ec76d5936e94abb651