meta-aspeed-sdk:aspeed-g7: Generate U-Boot FIT header only when UBOOT_FITIMAGE_ENABLE is set
Previously, the recovery UART recipe always generated and deployed the
U-Boot FIT header (u-boot-fit-header.bin) for the AST2700 A1 I2C/I3C recovery
image, even when the build did not use a FIT image.

This change adds a condition to generate and deploy the U-Boot FIT header
only when UBOOT_FITIMAGE_ENABLE is enabled. If this flag is set, the build
uses a U-Boot FIT image instead of a SoC manifest image, and the FIT header
must be included.

This improves build correctness and avoids generating unnecessary artifacts
for non-FIT image builds.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: Ie0b3259e16b326ea067c8e0fbbf642783be62030